==> scRNA_seq/R/00_genes_of_interest.R <==
integrins <- c(
  "ITGA1", "ITGA2", "ITGA3", "ITGA4", "ITGA5", "ITGA6", "ITGA7", "ITGA8", "ITGA9", "ITGA10", "ITGA11", "ITGAD", "ITGAE", "ITGAL",
  "ITGAM", "ITGAV", "ITGA2B", "ITGAX", "ITGB1", "ITGB2", "ITGB3", "ITGB4", "ITGB5", "ITGB6", "ITGB7", "ITGB8"
)
selectins <- c("SELE", "SELL", "SELP")
cadherins <- c("CDH1", "CDH2", "CDH12", "CDH3")
adhesion <- c(
  "NCAM1", "ICAM1", "VCAM1", "PECAM1", "MADCAM1", "CHL1", "NFASC", "NRCAM", "SIGLEC1",
  "CD22", "CD33", "MAG", "SIGLEC5", "SIGLEC6", "SIGLEC7", "SIGLEC10", "SIGLEC12",
  "NECTIN1", "NECTIN2", "NECTIN3", "NECTIN4",
  "CADM3", "CADM1", "CADM2", "CADM4", "PVR", "CD2", "CD48", "HEPACAM2", "DSCAM", "CD38"
)

HLA.I <- c("HLA-A", "HLA-B", "HLA-C", "HLA-E", "HLA-F", "HLA-G")
HLA.II <- c(
  "HLA-DMA", "HLA-DMB", "HLA-DOA", "HLA-DOB", "HLA-DPA1", "HLA-DPB1", "HLA-DQA1", "HLA-DQB1", "HLA-DQB1-AS1",
  "HLA-DQA2", "HLA-DQB2", "HLA-DRA", "HLA-DRB1", "HLA-DRB5"
)

checkpoint.molecules <- c("CD274", "PDCD1LG2", "CD80", "CD86", "LGALS9", "HMGB1", "CEACAM1", "NECTIN2", "PVR", "TNFRSF14", "CD47")
